Add "Browser default" option to "Close Tab button" preference
If this option set then Close Tab button will not be affected by this extension.

Introduce a new value of closeButtons preference - 10. Make it a default value.

Alias original `adjustTabstrip` function as `_originalAdjustTabstripFunc`
and override original function and call orginal aliased function in it.
It allows to add a condition (this.mCloseButtons == 10) when the original
function should be called.